services/mongo/requirements.txt
# Framework dependency
osprey-framework>=0.8.0

scipy
pymongo